+++ src/components/ExcelDataViewer.tsx
@@ -5,6 +5,12 @@ import { Input } from "@/components/ui/input";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Button } from "@/components/ui/button";
 import { Search, XCircle, FileSpreadsheet, Table as TableIcon } from "lucide-react";
+import {
+  columnLetterToIndex,
+  indexToColumnLetter,
+  extractColumnAndRow,
+  compareColumns,
+} from "@/lib/excelUtils";
 
 interface ExcelDataViewerProps {
   data: any;
@@ -32,48 +38,7 @@ const ExcelDataViewer: React.FC<ExcelDataViewerProps> = ({ data, title = "Datos 
   // Determina si los datos son un array o un objeto (formato de hoja de cálculo)
   const isArrayFormat = Array.isArray(data);
   
-  // Función para convertir letra de columna a índice numérico
-  const columnLetterToIndex = (columnLetter: string): number => {
-    let result = 0;
-    for (let i = 0; i < columnLetter.length; i++) {
-      result = result * 26 + columnLetter.charCodeAt(i) - 64;
-    }
-    return result;
-  };
-
-  // Función para convertir índice numérico a letra de columna
-  const indexToColumnLetter = (index: number): string => {
-    let columnLetter = '';
-    while (index > 0) {
-      const remainder = (index - 1) % 26;
-      columnLetter = String.fromCharCode(65 + remainder) + columnLetter;
-      index = Math.floor((index - 1) / 26);
-    }
-    return columnLetter || 'A';
-  };
-  
-  // Función para extraer columna y fila de un cellId
-  const extractColumnAndRow = (cellId: string): { column: string, row: number } | null => {
-    // Detecta columnas como A, B, C... Z, AA, AB, etc.
-    const match = cellId.match(/^([A-Z]+)(\d+)$/);
-    if (!match) return null;
-    
-    return {
-      column: match[1],
-      row: parseInt(match[2], 10)
-    };
-  };
-  
-  // Función para comparar columnas como A, B, C... Z, AA, AB, etc.
-  const compareColumns = (a: string, b: string): number => {
-    // Si son de longitud diferente, las columnas más largas van después (AA > Z)
-    if (a.length !== b.length) {
-      return a.length - b.length;
-    }
-    
-    // Si son de la misma longitud, comparar lexicográficamente
-    return a.localeCompare(b);
-  };
+  // Utils moved to '@/lib/excelUtils'
   
   // Para datos en formato de array (como en los logs)
   const renderArrayData = () => {
