+++ src/data/refrigerantTypes.ts
@@ -1,0 +1,15 @@
+export interface RefrigeranteData {
+  nombre: string;
+  composicion: string;
+  inflamabilidad: string;
+  toxicidad: string;
+  grupoSeguridad: string;
+  directivaEquipos: string;
+  pca: string;
+  agotamientoOzono: string;
+  limitePractico: string;
+  atelOdl: string;
+  limiteInflamabilidad: string;
+  temperaturaAutoignicion: string;
+  gasFluorado: string;
+}
