+++ src/data/refrigerantsData.part1.ts
@@ -1,0 +1,229 @@
+import type { RefrigeranteData } from "./refrigerantTypes";
+// Define refrigerants database
+export const refrigerantesDataPart1: Record<string, RefrigeranteData> = {
+  "R-1234yf": {
+    nombre: "R-1234yf",
+    composicion: "2,3,3,3-Tetrafluorpropeno - CF3CF=CH2",
+    inflamabilidad: "Grupo 2L",
+    toxicidad: "Grupo A",
+    grupoSeguridad: "A2L",
+    directivaEquipos: "1",
+    pca: "4",
+    agotamientoOzono: "0",
+    limitePractico: "0.058 kg/m3",
+    atelOdl: "0.47 kg/m3",
+    limiteInflamabilidad: "0.289 kg/m3",
+    temperaturaAutoignicion: "405 ºC",
+    gasFluorado: "NO"
+  },
+  "R-1234ze": {
+    nombre: "R-1234ze",
+    composicion: "Trans 1,3,3,3-Tetrafluorpropeno - CF3CH=CHF",
+    inflamabilidad: "Grupo 2L",
+    toxicidad: "Grupo A",
+    grupoSeguridad: "A2L",
+    directivaEquipos: "2",
+    pca: "7",
+    agotamientoOzono: "0",
+    limitePractico: "0.061 kg/m3",
+    atelOdl: "0.28 kg/m3",
+    limiteInflamabilidad: "0.303 kg/m3",
+    temperaturaAutoignicion: "368 ºC",
+    gasFluorado: "NO"
+  },
+  "R-134a": {
+    nombre: "R-134a",
+    composicion: "1,1,1,2-Tetrafluoretano - CF3CH2F",
+    inflamabilidad: "Grupo 1",
+    toxicidad: "Grupo A",
+    grupoSeguridad: "A1",
+    directivaEquipos: "2",
+    pca: "1.430",
+    agotamientoOzono: "0",
+    limitePractico: "0.25 kg/m3",
+    atelOdl: "0.21 kg/m3",
+    limiteInflamabilidad: "NF",
+    temperaturaAutoignicion: "743 ºC",
+    gasFluorado: "SI"
+  },
+  "R-170": {
+    nombre: "R-170",
+    composicion: "Etano - C2H6",
+    inflamabilidad: "Grupo 3",
+    toxicidad: "Grupo A",
+    grupoSeguridad: "A3",
+    directivaEquipos: "1",
+    pca: "6",
+    agotamientoOzono: "0",
+    limitePractico: "0.0086 kg/m3",
+    atelOdl: "0.0086 kg/m3",
+    limiteInflamabilidad: "0.038 kg/m3",
+    temperaturaAutoignicion: "515 ºC",
+    gasFluorado: "NO"
+  },
+  "R-22": {
+    nombre: "R-22",
+    composicion: "Clorodifluormetano - CHClF2",
+    inflamabilidad: "Grupo 1",
+    toxicidad: "Grupo A",
+    grupoSeguridad: "A1",
+    directivaEquipos: "2",
+    pca: "1.810",
+    agotamientoOzono: "0.055",
+    limitePractico: "0.3 kg/m3",
+    atelOdl: "0.21 kg/m3",
+    limiteInflamabilidad: "NF",
+    temperaturaAutoignicion: "635 ºC",
+    gasFluorado: "SI"
+  },
+  "R-290": {
+    nombre: "R-290",
+    composicion: "Propano - C3H8",
+    inflamabilidad: "Grupo 3",
+    toxicidad: "Grupo A",
+    grupoSeguridad: "A3",
+    directivaEquipos: "1",
+    pca: "3",
+    agotamientoOzono: "0",
+    limitePractico: "0.008 kg/m3",
+    atelOdl: "0.09 kg/m3",
+    limiteInflamabilidad: "0.038 kg/m3",
+    temperaturaAutoignicion: "470 ºC",
+    gasFluorado: "NO"
+  },
+  "R-32": {
+    nombre: "R-32",
+    composicion: "Difluormetano - CH2F2",
+    inflamabilidad: "Grupo 2L",
+    toxicidad: "Grupo A",
+    grupoSeguridad: "A2L",
+    directivaEquipos: "1",
+    pca: "675",
+    agotamientoOzono: "0",
+    limitePractico: "0.061 kg/m3",
+    atelOdl: "0.3 kg/m3",
+    limiteInflamabilidad: "0.307 kg/m3",
+    temperaturaAutoignicion: "648 ºC",
+    gasFluorado: "SI"
+  },
+  "R-404A": {
+    nombre: "R-404A",
+    composicion: "44% R-125 / 52% R-143a / 4% R-134a",
+    inflamabilidad: "Grupo 1",
+    toxicidad: "Grupo A",
+    grupoSeguridad: "A1",
+    directivaEquipos: "2",
+    pca: "3.922",
+    agotamientoOzono: "0",
+    limitePractico: "0.52 kg/m3",
+    atelOdl: "0.52 kg/m3",
+    limiteInflamabilidad: "NF",
+    temperaturaAutoignicion: "728 ºC",
+    gasFluorado: "SI"
+  },
+  "R-407F": {
+    nombre: "R-407F",
+    composicion: "30% R-32 / 30% R-125 / 40% R-134a",
+    inflamabilidad: "Grupo 1",
+    toxicidad: "Grupo A",
+    grupoSeguridad: "A1",
+    directivaEquipos: "2",
+    pca: "1.825",
+    agotamientoOzono: "0",
+    limitePractico: "0.32 kg/m3",
+    atelOdl: "0.32 kg/m3",
+    limiteInflamabilidad: "NF",
+    temperaturaAutoignicion: "ND",
+    gasFluorado: "SI"
+  },
+  "R-410A": {
+    nombre: "R-410A",
+    composicion: "50% R-32 / 50% R-125",
+    inflamabilidad: "Grupo 1",
+    toxicidad: "Grupo A",
+    grupoSeguridad: "A1",
+    directivaEquipos: "2",
+    pca: "2.088",
+    agotamientoOzono: "0",
+    limitePractico: "0.44 kg/m3",
+    atelOdl: "0.42 kg/m3",
+    limiteInflamabilidad: "NF",
+    temperaturaAutoignicion: "ND",
+    gasFluorado: "SI"
+  },
+  "R-424A": {
+    nombre: "R-424A",
+    composicion: "50,5% R-125 / 47% R-134a / 0,9% R-600a / 1% R-600 / 0,6% R-601a",
+    inflamabilidad: "Grupo 1",
+    toxicidad: "Grupo A",
+    grupoSeguridad: "A1",
+    directivaEquipos: "2",
+    pca: "2.440",
+    agotamientoOzono: "0",
+    limitePractico: "0.1 kg/m3",
+    atelOdl: "0.1 kg/m3",
+    limiteInflamabilidad: "NF",
+    temperaturaAutoignicion: "ND",
+    gasFluorado: "SI"
+  },
+  "R-427A": {
+    nombre: "R-427A",
+    composicion: "15% R-32 / 25% R-125 / 10% R-143a / 50% R-134a",
+    inflamabilidad: "Grupo 1",
+    toxicidad: "Grupo A",
+    grupoSeguridad: "A1",
+    directivaEquipos: "2",
+    pca: "2.138",
+    agotamientoOzono: "0",
+    limitePractico: "0.29 kg/m3",
+    atelOdl: "0.29 kg/m3",
+    limiteInflamabilidad: "NF",
+    temperaturaAutoignicion: "ND",
+    gasFluorado: "SI"
+  },
+  "R-434A": {
+    nombre: "R-434A",
+    composicion: "63,2% R-125 / 18% R-143a / 16% R-134a / 2,8% R-600a",
+    inflamabilidad: "Grupo 1",
+    toxicidad: "Grupo A",
+    grupoSeguridad: "A1",
+    directivaEquipos: "2",
+    pca: "3.245",
+    agotamientoOzono: "0",
+    limitePractico: "0.32 kg/m3",
+    atelOdl: "0.32 kg/m3",
+    limiteInflamabilidad: "NF",
+    temperaturaAutoignicion: "ND",
+    gasFluorado: "SI"
+  },
+  "R-442A": {
+    nombre: "R-442A",
+    composicion: "31% R-32 / 31% R-125 / 30% R-134a / 3% R-152a / 5% R-227ea",
+    inflamabilidad: "Grupo 1",
+    toxicidad: "Grupo A",
+    grupoSeguridad: "A1",
+    directivaEquipos: "2",
+    pca: "1.888",
+    agotamientoOzono: "0",
+    limitePractico: "0.33 kg/m3",
+    atelOdl: "0.33 kg/m3",
+    limiteInflamabilidad: "NF",
+    temperaturaAutoignicion: "ND",
+    gasFluorado: "SI"
+  },
+  "R-448A": {
+    nombre: "R-448A",
+    composicion: "26% R-32 / 26% R-125 / 20% R-1234yf / 21% R-134a / 7% R-1234ze",
+    inflamabilidad: "Grupo 1",
+    toxicidad: "Grupo A",
+    grupoSeguridad: "A1",
+    directivaEquipos: "2",
+    pca: "1.387",
+    agotamientoOzono: "0",
+    limitePractico: "0.388 kg/m3",
+    atelOdl: "0.388 kg/m3",
+    limiteInflamabilidad: "NF",
+    temperaturaAutoignicion: "ND",
+    gasFluorado: "SI"
+  },
+};
