+++ src/data/refrigerantsData.part2.ts
@@ -1,0 +1,183 @@
+import type { RefrigeranteData } from "./refrigerantTypes";
+export const refrigerantesDataPart2: Record<string, RefrigeranteData> = {
+  "R-449A": {
+    nombre: "R-449A",
+    composicion: "24,3% R-32 / 24,7% R-125 / 25,3% R-1234yf / 25,7% R-134a",
+    inflamabilidad: "Grupo 1",
+    toxicidad: "Grupo A",
+    grupoSeguridad: "A1",
+    directivaEquipos: "2",
+    pca: "1.397",
+    agotamientoOzono: "0",
+    limitePractico: "0.357 kg/m3",
+    atelOdl: "0.357 kg/m3",
+    limiteInflamabilidad: "NF",
+    temperaturaAutoignicion: "ND",
+    gasFluorado: "SI"
+  },
+  "R-450A": {
+    nombre: "R-450A",
+    composicion: "42% R-134a / 58% R-1234ze",
+    inflamabilidad: "Grupo 1",
+    toxicidad: "Grupo A",
+    grupoSeguridad: "A1",
+    directivaEquipos: "2",
+    pca: "604.7",
+    agotamientoOzono: "0",
+    limitePractico: "0.319 kg/m3",
+    atelOdl: "0.319 kg/m3",
+    limiteInflamabilidad: "NF",
+    temperaturaAutoignicion: "ND",
+    gasFluorado: "SI"
+  },
+  "R-452A": {
+    nombre: "R-452A",
+    composicion: "11% R-32 / 59% R-125 / 30% R-1234yf",
+    inflamabilidad: "Grupo 1",
+    toxicidad: "Grupo A",
+    grupoSeguridad: "A1",
+    directivaEquipos: "2",
+    pca: "2.140",
+    agotamientoOzono: "0",
+    limitePractico: "0.423 kg/m3",
+    atelOdl: "0.423 kg/m3",
+    limiteInflamabilidad: "NF",
+    temperaturaAutoignicion: "ND",
+    gasFluorado: "SI"
+  },
+  "R-453A": {
+    nombre: "R-453A",
+    composicion: "20% R-32 / 20% R-125 / 53,8% R-134a / 5% R-227ea / 0,6% R-600 / 0,6% R-601",
+    inflamabilidad: "Grupo 1",
+    toxicidad: "Grupo A",
+    grupoSeguridad: "A1",
+    directivaEquipos: "2",
+    pca: "1.765.4",
+    agotamientoOzono: "0",
+    limitePractico: "0.14 kg/m3",
+    atelOdl: "0.14 kg/m3",
+    limiteInflamabilidad: "NF",
+    temperaturaAutoignicion: "ND",
+    gasFluorado: "SI"
+  },
+  "R-454C": {
+    nombre: "R-454C",
+    composicion: "21,5% R-32 / 78,5% R-1234yf",
+    inflamabilidad: "Grupo 2L",
+    toxicidad: "Grupo A",
+    grupoSeguridad: "A2L",
+    directivaEquipos: "1",
+    pca: "148.27",
+    agotamientoOzono: "0",
+    limitePractico: "0.059 kg/m3",
+    atelOdl: "0.44 kg/m3",
+    limiteInflamabilidad: "0.291 kg/m3",
+    temperaturaAutoignicion: "444 ºC",
+    gasFluorado: "SI"
+  },
+  "R-455A": {
+    nombre: "R-455A",
+    composicion: "3% R-744 / 21,5% R-32 / 75,5% R-1234yf",
+    inflamabilidad: "Grupo 2L",
+    toxicidad: "Grupo A",
+    grupoSeguridad: "A2L",
+    directivaEquipos: "1",
+    pca: "148.18",
+    agotamientoOzono: "0",
+    limitePractico: "0.105 kg/m3",
+    atelOdl: "0.414 kg/m3",
+    limiteInflamabilidad: "0.423 kg/m3",
+    temperaturaAutoignicion: "473 ºC",
+    gasFluorado: "SI"
+  },
+  "R-507A": {
+    nombre: "R-507A",
+    composicion: "50% R-125 / 50% R-143a",
+    inflamabilidad: "Grupo 1",
+    toxicidad: "Grupo A",
+    grupoSeguridad: "A1",
+    directivaEquipos: "2",
+    pca: "3.985",
+    agotamientoOzono: "0",
+    limitePractico: "0.53 kg/m3",
+    atelOdl: "0.53 kg/m3",
+    limiteInflamabilidad: "NF",
+    temperaturaAutoignicion: "ND",
+    gasFluorado: "SI"
+  },
+  "R-513A": {
+    nombre: "R-513A",
+    composicion: "44% R-134a / 56% R-1234yf",
+    inflamabilidad: "Grupo 1",
+    toxicidad: "Grupo A",
+    grupoSeguridad: "A1",
+    directivaEquipos: "2",
+    pca: "631.4",
+    agotamientoOzono: "0",
+    limitePractico: "0.319 kg/m3",
+    atelOdl: "0.319 kg/m3",
+    limiteInflamabilidad: "NF",
+    temperaturaAutoignicion: "ND",
+    gasFluorado: "SI"
+  },
+  "R-600": {
+    nombre: "R-600",
+    composicion: "Butano - C4H10",
+    inflamabilidad: "Grupo 3",
+    toxicidad: "Grupo A",
+    grupoSeguridad: "A3",
+    directivaEquipos: "1",
+    pca: "4",
+    agotamientoOzono: "0",
+    limitePractico: "0.0089 kg/m3",
+    atelOdl: "0.0024 kg/m3",
+    limiteInflamabilidad: "0.038 kg/m3",
+    temperaturaAutoignicion: "365 ºC",
+    gasFluorado: "NO"
+  },
+  "R-600a": {
+    nombre: "R-600a",
+    composicion: "Isobutano - CH(CH3)3",
+    inflamabilidad: "Grupo 3",
+    toxicidad: "Grupo A",
+    grupoSeguridad: "A3",
+    directivaEquipos: "1",
+    pca: "3",
+    agotamientoOzono: "0",
+    limitePractico: "0.011 kg/m3",
+    atelOdl: "0.059 kg/m3",
+    limiteInflamabilidad: "0.043 kg/m3",
+    temperaturaAutoignicion: "460 ºC",
+    gasFluorado: "NO"
+  },
+  "R-717": {
+    nombre: "R-717",
+    composicion: "Amoniaco - NH3",
+    inflamabilidad: "Grupo 2L",
+    toxicidad: "Grupo B",
+    grupoSeguridad: "B2L",
+    directivaEquipos: "1",
+    pca: "0",
+    agotamientoOzono: "0",
+    limitePractico: "0.00035 kg/m3",
+    atelOdl: "0.00022 kg/m3",
+    limiteInflamabilidad: "0.116 kg/m3",
+    temperaturaAutoignicion: "630 ºC",
+    gasFluorado: "NO"
+  },
+  "R-744": {
+    nombre: "R-744",
+    composicion: "Dióxido de carbono - CO2",
+    inflamabilidad: "Grupo 1",
+    toxicidad: "Grupo A",
+    grupoSeguridad: "A1",
+    directivaEquipos: "2",
+    pca: "1",
+    agotamientoOzono: "0",
+    limitePractico: "0.1 kg/m3",
+    atelOdl: "0.072 kg/m3",
+    limiteInflamabilidad: "NF",
+    temperaturaAutoignicion: "ND",
+    gasFluorado: "NO"
+  }
+};
